backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
python-dotenv
psycopg[binary]
